+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/utils/tf/loaders/BroadcastGradientArgs.scala
@@ -19,7 +19,7 @@ import java.nio.ByteOrder
 
 import com.intel.analytics.bigdl.Module
 import com.intel.analytics.bigdl.nn.Identity
-import com.intel.analytics.bigdl.nn.ops.BroadcastGradientArgs
+import com.intel.analytics.bigdl.nn.tf.{BroadcastGradientArgs => BroadcastGradientArgsOps}
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.tf.Context
@@ -33,6 +33,6 @@ class BroadcastGradientArgs extends TensorflowOpsLoader {
 
   override def build[T: ClassTag](nodeDef: NodeDef, byteOrder: ByteOrder,
     context: Context[T])(implicit ev: TensorNumeric[T]): Module[T] = {
-    BroadcastGradientArgs[T]()
+    new BroadcastGradientArgsOps[T]()
   }
 }
